added unit tests and rearranged code structure to include a database.py file
