+++ tests/test_console.py
@@ -1,3 +1,5 @@
+import json
+import os
 import requests
 import server
 from time import sleep
@@ -8,6 +10,7 @@ from testfixtures import log_capture
 import testing.postgresql
 
 from console import views
+from console.database import SurveyResponse
 from console.helpers.exceptions import ClientError, ServiceError
 
 
@@ -123,3 +126,78 @@ class TestAuthentication(unittest.TestCase):
         self.login('none', 'password')
         response = self.app.get('/admin/user', follow_redirects=True)
         self.assertEqual(403, response.status_code)
+
+
+class TestStore(unittest.TestCase):
+
+    def setUp(self):
+        self.postgres = Postgresql()
+        Postgresql.clear_cache()
+        self.app = server.app.test_client()
+        self.app.testing = True
+        self.render_templates = False
+        TestAuthentication.login(self, 'dev', 'password')
+
+    def get_test_responses(self):
+        SITE_ROOT = os.path.realpath(os.path.dirname(__file__))
+        json_url = os.path.join(SITE_ROOT, 'test_data', 'test_response_1.json')
+        responses_json = []
+        with open(json_url) as json_data:
+            responses_json.append(json.load(json_data))
+        survey_responses = []
+        for response in responses_json:
+            tx_id = response['tx_id']
+            invalid = 'f'
+            data = response
+            survey_responses.append(SurveyResponse(tx_id=tx_id,
+                                                   invalid=invalid,
+                                                   data=data))
+        return survey_responses
+
+    def test_display_data(self):
+        responses = self.get_test_responses()
+        with mock.patch.object(views, 'get_all_responses', return_value=responses):
+            response = self.app.get('/store', follow_redirects=True)
+            self.assertIn(b'f088d89d-a367-876e-f29f-ae8f1a260000', response.data)
+
+    def test_tx_id_search_success(self):
+        responses = self.get_test_responses()
+        with mock.patch.object(views, 'get_all_responses', return_value=responses):
+            response = self.app.get('/store?tx_id=f088d89d-a367-876e-f29f-ae8f1a260000',
+                                    follow_redirects=True)
+            self.assertIn(b'f088d89d-a367-876e-f29f-ae8f1a260000', response.data)
+
+    def test_tx_id_search_fail(self):
+        responses = self.get_test_responses()
+        with mock.patch.object(views, 'get_all_responses', return_value=responses):
+            response = self.app.get('/store?tx_id=f088d89d-a367-876e-f29f-ae8f1a260001',
+                                    follow_redirects=True)
+            self.assertNotIn(b'f088d89d-a367-876e-f29f-ae8f1a260000', response.data)
+
+    def test_ru_ref_search_success(self):
+        responses = self.get_test_responses()
+        with mock.patch.object(views, 'get_all_responses', return_value=responses):
+            response = self.app.get('/store?ru_ref=12345678901a',
+                                    follow_redirects=True)
+            self.assertIn(b'f088d89d-a367-876e-f29f-ae8f1a260000', response.data)
+
+    def test_ru_ref_search_fail(self):
+        responses = self.get_test_responses()
+        with mock.patch.object(views, 'get_all_responses', return_value=responses):
+            response = self.app.get('/store?ru_ref=12345678901f',
+                                    follow_redirects=True)
+            self.assertNotIn(b'f088d89d-a367-876e-f29f-ae8f1a260000', response.data)
+
+    def test_datetime_search_success(self):
+        responses = self.get_test_responses()
+        with mock.patch.object(views, 'get_all_responses', return_value=responses):
+            response = self.app.get('/store?datetime_earliest=2015-01-01T01%3A01&datetime_latest=2020-01-01T01%3A01',
+                                    follow_redirects=True)
+            self.assertIn(b'f088d89d-a367-876e-f29f-ae8f1a260000', response.data)
+
+    def test_datetime_search_fail(self):
+        responses = self.get_test_responses()
+        with mock.patch.object(views, 'get_all_responses', return_value=responses):
+            response = self.app.get('/store?datetime_earliest=2020-01-01T01%3A01&datetime_latest=2020-01-01T01%3A01',
+                                    follow_redirects=True)
+            self.assertNotIn(b'f088d89d-a367-876e-f29f-ae8f1a260000', response.data)
